Change the popular sidebar categories area of the skin to use valid HTML.
This commit introduces a change to the Answers skin. It fixes several
basic HTML validation errors while maintaining the two-column structure
and other stylistic components of the popular categories in the sidebar.

IMPORTANT: This commit requires that the contents of
[[MediaWiki:Sidebar-popular-categories]] be changed to something like the
following HTML code. Note that `<ul>` elements are manually entered, and
that the wrapping `<div>` elements and spacing `<br />` elements that used
to be present must be removed.

USE HTML LIKE THIS IN THE MEDIAWIKI LIST MESSAGE WHEN THIS CHANGE IS RELEASED:

<table align='center' width='100%'><tr><td width='50%' valign='top'>
<ul>
* Category:Relationships|<b><big>Relationships</big></b>
* Category:Pregnancy|Pregnancy
* Category:Sex|Sex
</ul><ul>
* Category:Money|<b><big>Money</big></b>
* Category:Finance|Finance
* Category:Jobs|Jobs
</ul>
</td>
<td width='50%' valign='top'>
<ul>
* Category:Cooking|<b><big>Cooking</big></b>
* Category:Recipes|Recipes
* Category:Food|Food
</ul><ul>
* Category:Puzzles|<b><big>Puzzles</big></b>
* Category:Crossword Puzzles|Crosswords
* Category:Words|Words
* Category:Letter equations|Letter equations
</ul>
</td></tr></table>



git-svn-id: https://svn.wikia-code.com/wikia/trunk@16516 e846bd91-c042-42a7-a218-4b65e2c1904d