BugzId 11047 - Making Chat rail module behave better during failover.  Did the following during read-only: - TECH FIX: Made all API calls to Node be skipped in wgReadOnly because it's possible the read-only is due to a complete network failure in the primary datacenter (kinda weird assumption, but we don't have any other wiki-vars to express this specific type of problem). This fix prevents any sort of hanging while waiting for responses from a server we're not going to be able to reach. - INTERFACE FIX: Made the entry-point not display the chatters and the chat button during failover, but rather it displays an i18n message that explains that Chat is not available in read-only mode.
Keep in mind that chat MAY work in read-only depending on the failure since it doesn't need to write to the production databases for anything such as promotions and kick-bans. By disabling the entry-point we're not advertising it as working... if it works & users go to it via the URL, then that's bonus for them (but they're playing in the Wild West w/no kickbans ;) ).  We may change our mind on whether we want to completely disable it, but I think our hardcore users will appreciate being able to chat during read-only if they are lucky enough that it still works during the particular failure.

NOTE: More testing of this pending... may be some followup commits.

git-svn-id: https://svn.wikia-code.com/wikia/trunk@41662 e846bd91-c042-42a7-a218-4b65e2c1904d